Add hw_18
Add hw_18

+++ homework_18/src/commercial/An124.java
@@ -2,7 +2,7 @@ package commercial;
 
 public class An124 extends CommercialAirplane {
 
-    public An124(int numberOfPassengers, int capacityOfShipment, int normalSpeed, float fuelConsumption, int volumeOfTank, int numOfPorters) {
+    public An124(int numberOfPassengers, int capacityOfShipment, int normalSpeed, int fuelConsumption, int volumeOfTank, int numOfPorters) {
         super(numberOfPassengers, capacityOfShipment, normalSpeed, fuelConsumption, volumeOfTank, numOfPorters);
     }
