+++ homework_18/src/commercial/Bristol170.java
@@ -2,7 +2,7 @@ package commercial;
 
 public class Bristol170 extends CommercialAirplane {
 
-    public Bristol170(int numberOfPassengers, int capacityOfShipment, int normalSpeed, float fuelConsumption, int volumeOfTank, int numOfPorters) {
+    public Bristol170(int numberOfPassengers, int capacityOfShipment, int normalSpeed, int fuelConsumption, int volumeOfTank, int numOfPorters) {
         super(numberOfPassengers, capacityOfShipment, normalSpeed, fuelConsumption, volumeOfTank, numOfPorters);
     }
