+++ homework_18/src/military/LockheedF117.java
@@ -2,7 +2,7 @@ package military;
 
 public class LockheedF117 extends MilitaryAirplane {
 
-    public LockheedF117(int numberOfPassengers, int capacityOfShipment, int normalSpeed, float fuelConsumption, int volumeOfTank, boolean isWeaponOnTheBoard) {
+    public LockheedF117(int numberOfPassengers, int capacityOfShipment, int normalSpeed, int fuelConsumption, int volumeOfTank, boolean isWeaponOnTheBoard) {
         super(numberOfPassengers, capacityOfShipment, normalSpeed, fuelConsumption, volumeOfTank, isWeaponOnTheBoard);
     }
