+++ homework_18/src/military/Su47.java
@@ -2,7 +2,7 @@ package military;
 
 public class Su47 extends MilitaryAirplane{
 
-    public Su47(int numberOfPassengers, int capacityOfShipment, int normalSpeed, float fuelConsumption, int volumeOfTank, boolean isWeaponOnTheBoard) {
+    public Su47(int numberOfPassengers, int capacityOfShipment, int normalSpeed, int fuelConsumption, int volumeOfTank, boolean isWeaponOnTheBoard) {
         super(numberOfPassengers, capacityOfShipment, normalSpeed, fuelConsumption, volumeOfTank, isWeaponOnTheBoard);
     }
