+++ homework_18/src/passengers/AirbusA340.java
@@ -2,7 +2,7 @@ package passengers;
 
 public class AirbusA340 extends PassengerAirplane {
 
-    public AirbusA340(int numberOfPassengers, int capacityOfShipment, int normalSpeed, float fuelConsumption, int volumeOfTank, int numOfStewardess) {
+    public AirbusA340(int numberOfPassengers, int capacityOfShipment, int normalSpeed, int fuelConsumption, int volumeOfTank, int numOfStewardess) {
         super(numberOfPassengers, capacityOfShipment, normalSpeed, fuelConsumption, volumeOfTank, numOfStewardess);
     }
