+++ homework_18/src/passengers/PassengerAirplane.java
@@ -1,10 +1,11 @@
 package passengers;
+import plane.Airplane;
 
 public abstract class PassengerAirplane extends Airplane {
 
     int numOfStewardess;
 
-    public PassengerAirplane(int numberOfPassengers, int capacityOfShipment, int normalSpeed, float fuelConsumption, int volumeOfTank, int numOfStewardess) {
+    public PassengerAirplane(int numberOfPassengers, int capacityOfShipment, int normalSpeed, int fuelConsumption, int volumeOfTank, int numOfStewardess) {
         super(numberOfPassengers, capacityOfShipment, normalSpeed, fuelConsumption, volumeOfTank);
         this.numOfStewardess = numOfStewardess;
     }
